requirements.txt
streamlit
scikit-learn
joblib
speechrecognition
pyaudio
geocoder
requests
python-dotenv
twilio
